extract Commit logic from trie to snapshot
